requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
pytest==7.3.2
requests==2.32.5
flask-login==0.6.3
